chore: polish code, add tests, and improve error handling
- add top-level crate documentation for nwr
- format SQLite in doc comments for db.rs
- replace deprecated map/unwrap_or with is_some_and for gz check
- reorganize module exports with doc comments
- fix MAX_TAXONOMY_DEPTH type mismatch in taxonomy code
- simplify load_file argument handling in seqdb subcommand
- simplify cache lookups in seqdb and taxonomy libraries
- move ASS_COLUMNS constant earlier in template subcommand
- add tests for empty input files and invalid taxon terms
- add test assembly summary with blank lines

+++ src/libs/taxonomy.rs
@@ -339,7 +339,7 @@ pub fn get_ancestor(conn: &rusqlite::Connection, id: i64) -> anyhow::Result<Taxo
 ///
 /// NCBI taxonomy depths are well below this limit; the bound acts as a guard
 /// against corrupt cyclic data.
-const MAX_TAXONOMY_DEPTH: usize = 1000;
+const MAX_TAXONOMY_DEPTH: i64 = 1000;
 
 /// All Nodes to the root (with ID 1)
 ///
@@ -377,7 +377,7 @@ pub fn get_lineage(conn: &rusqlite::Connection, id: i64) -> anyhow::Result<Vec<T
         ",
     )?;
 
-    let mut rows = stmt.query(rusqlite::params![id, MAX_TAXONOMY_DEPTH as i64])?;
+    let mut rows = stmt.query(rusqlite::params![id, MAX_TAXONOMY_DEPTH])?;
     let mut ids = Vec::new();
     let mut seen = std::collections::HashSet::new();
 
@@ -491,7 +491,7 @@ pub fn get_all_descendent(
         ",
     )?;
 
-    let mut rows = stmt.query(rusqlite::params![id, MAX_TAXONOMY_DEPTH as i64])?;
+    let mut rows = stmt.query(rusqlite::params![id, MAX_TAXONOMY_DEPTH])?;
     let mut ids = Vec::new();
     let mut seen = std::collections::HashSet::new();
